help the compiler create more efficient code
Patch from Michael <virtual_worlds@gmx.de>
Full text: As usual I have modified only minor things to remove warnigs when compiled as C++ code, have added some const-specifiers where possible to help the compiler create more efficient code and added some static-specifiers to make functions invisible to other modules.

+++ backend/auspost.c
@@ -69,7 +69,7 @@ static inline char convert_pattern(char data, int shift) {
 
 /* Adds Reed-Solomon error correction to auspost */
 void rs_error(char data_pattern[]) {
-    int reader, triple_writer = 0;
+    size_t reader, triple_writer = 0;
     char triple[31], inv_triple[31];
     unsigned char result[5];
